requirements.txt
pdf2image
Pillow
streamlit
googletrans
speechrecognition
gtts
transformers
fpdf
easyocr
pydub
PyPdf2
